# storage/api-docs/api/v1/doc/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Welcome to PayLibero's APIs Panel!",
        "description": "This is a documentation covers all the available services at the backend of the project. For further information about the parameters, please feel free to explore  the Schema part of each API. The Schema part is available for POST and PUT HTTP services.",
        "version": "0.1"
    },
    "paths": {
...
        "/api/applicant/signin": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign in an applicant",
                "operationId": "applicantSignin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "An encrypted token for authorizing the signed in user.",
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "A unique identifier for the user in the dataset.",
                                                        "type": "integer"
                                                    },
                                                    "role": {
                                                        "description": "A field indicates user role. 0: Applicant, 1: Admin.",
                                                        "type": "integer"
                                                    },
                                                    "first_name": {
                                                        "description": "A field indicates user's first name.",
                                                        "type": "string"
                                                    },
                                                    "last_name": {
                                                        "description": "A field indicates user's last name.",
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "description": "A field indicates user's email.",
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "description": "A field indicates user's status. 0: Deactive, 1: Active",
                                                        "type": "integer"
                                                    },
                                                    "is_email_verified": {
                                                        "description": "A field indicates whther user's email is verified or not. true: is_verified, false: is_not_verified",
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/applicant/signout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign out an applicant",
                "operationId": "applicantSignout",
                "responses": {
                    "200": {
                        "description": "Successful operation."
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applicant/set-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Set password to the applicant account by authenticated applicant",
                "operationId": "setPasswordToApplicantAccountByApplicant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "The password must contain at least 8 digits, one lowercase letter, one uppercase letter, one digit, and one special character",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "A descriptive attribute indicating the result of request.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applicant/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset the password of the applicant account by authenticated applicant",
                "operationId": "resetPasswordOfApplicantAccountByApplicant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "The new password must contain at least 8 digits, one lowercase letter, one uppercase letter, one digit, and one special character",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "A descriptive attribute indicating the result of request.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applicant/get-token": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get token included into cookie",
                "operationId": "getTokenIncludedInCookieByApplicant",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "An access token dispatched from the cookie.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}